[INLONG-3202][SDK] Unify SortSourceConfig of manager and sdk
