+++ inlong-common/src/main/java/org/apache/inlong/common/pojo/sdk/CacheZone.java
@@ -15,50 +15,25 @@
  * limitations under the License.
  */
 
-package org.apache.inlong.manager.common.pojo.sort;
+package org.apache.inlong.common.pojo.sdk;
 
-import io.swagger.annotations.ApiModel;
+import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
+import lombok.NoArgsConstructor;
 
 import java.util.List;
 import java.util.Map;
 
 @Data
 @Builder
-@ApiModel("Sort source sdk config")
-public class SortSourceConfigResponse {
-    String msg;
-    int code;
-    String md5;
-    SortSourceConfig data;
-
-    @Data
-    @Builder
-    public static class SortSourceConfig {
-        String sortClusterName;
-        String sortTaskId;
-        Map<String, CacheZone> cacheZones;
-    }
-
-    @Data
-    @Builder
-    public static class CacheZone {
-        String zoneName;
-        String serviceUrl;
-        String authentication;
-        List<Topic> topics;
-        Map<String, String> cacheZoneProperties;
-        String zoneType;
-    }
-
-    @Data
-    @Builder
-    public static class Topic {
-        String topic;
-        int partitionCnt;
-        Map<String, String> topicProperties;
-    }
+@AllArgsConstructor
+@NoArgsConstructor
+public class CacheZone {
+    String zoneName;
+    String serviceUrl;
+    String authentication;
+    List<Topic> topics;
+    Map<String, String> cacheZoneProperties;
+    String zoneType;
 }
-
-
