+++ inlong-common/src/main/java/org/apache/inlong/common/pojo/sdk/SortSourceConfigResponse.java
@@ -15,22 +15,23 @@
  * limitations under the License.
  */
 
-package org.apache.inlong.manager.service.core;
+package org.apache.inlong.common.pojo.sdk;
 
-import org.apache.inlong.common.pojo.sdk.CacheZone;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
-import java.util.Map;
+@Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class SortSourceConfigResponse {
+    String msg;
+    int code;
+    String md5;
+    CacheZoneConfig data;
 
-/**
- * Sort source service.
- */
-public interface SortSourceService {
-
-    /**
-     * Get cache zones by cluster name and task name.
-     * @param clusterName Target cluster name.
-     * @param taskName Target task name.
-     * @return SortSourceConfigResponse
-     */
-    Map<String, CacheZone> getCacheZones(String clusterName, String taskName);
 }
+
+
