+++ inlong-common/src/main/java/org/apache/inlong/common/pojo/sdk/Topic.java
@@ -15,22 +15,21 @@
  * limitations under the License.
  */
 
-package org.apache.inlong.manager.service.core;
+package org.apache.inlong.common.pojo.sdk;
 
-import org.apache.inlong.common.pojo.sdk.CacheZone;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
 import java.util.Map;
 
-/**
- * Sort source service.
- */
-public interface SortSourceService {
-
-    /**
-     * Get cache zones by cluster name and task name.
-     * @param clusterName Target cluster name.
-     * @param taskName Target task name.
-     * @return SortSourceConfigResponse
-     */
-    Map<String, CacheZone> getCacheZones(String clusterName, String taskName);
+@Data
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor
+public class Topic {
+    String topic;
+    int partitionCnt;
+    Map<String, String> topicProperties;
 }
